source/application/backend/requirements.txt
keras==2.4.3
tensorflow>=2.3.0
numpy>=1.16.6
flask>=1.0.2
Flask-Cors>=3.0.9
pillow>=6.2.0
h5py>=2.8.0
matplotlib>=2.2.5
pymemcache>=3.5.0